+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Use `cargo release` to create a new release.
 
 ## [Unreleased]
 
+## [2.0.3] – 2023-04-24
+
 ### Changed
 - mdcat now uses the kitty protocol to render images on WezTerm (see [GH-258]).
 - mdcat now downscales images to the column limit if rendering with the kitty protocol (see [GH-258]).
@@ -698,7 +700,8 @@ Use `cargo release` to create a new release.
 - Support ordered and unordered lists, with nest.
 - Show links, with references grouped by section.
 
-[Unreleased]: https://github.com/swsnr/mdcat/compare/mdcat-2.0.2...HEAD
+[Unreleased]: https://github.com/swsnr/mdcat/compare/mdcat-2.0.3...HEAD
+[2.0.3]: https://github.com/swsnr/mdcat/compare/mdcat-2.0.2...mdcat-2.0.3
 [2.0.2]: https://github.com/swsnr/mdcat/compare/mdcat-2.0.1...mdcat-2.0.2
 [2.0.1]: https://github.com/swsnr/mdcat/compare/mdcat-2.0.0...mdcat-2.0.1
 [2.0.0]: https://github.com/swsnr/mdcat/compare/mdcat-1.1.1...mdcat-2.0.0